services/eventos/requirements.txt
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.7.4
charset-normalizer==3.3.2
cryptography==42.0.8
ecdsa==0.18.0
fastapi==0.111.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1